added additional documentation for convolve and lfilter, renamed functional_filtering to functional_sox_convenience
